Create FLPN2lawyer_list.yml
This is the list of lawyers, arranged in alphabetical order according to surname.

This is a starting point. We need to decide what info will be used to categorise the lawyers, and what info will simply be for display purposes.

The information here is mostly the raw answers that lawyers have provided on the survey. Most will need to be rewritten.

note: some of the lawyers misunderstood the question "what areas do you service", and answered with areas of laws rather than areas in Adelaide. I figured we can find this information on their website.